[refactor] [#1] 고급설정 사이드바 컴포넌트 배치 방식 변경
- 고급 설정 사이드바에서 세부 유형 변경 시 기능 유형 컴포넌트까지 리렌더링이 되는 현상발생
- 세부유형 컴포넌트를 기능유형 컴포넌트 하위에 넣음
- 스타일 텀포넌트를 세부유형 컴포넌트 하위에 넣음
