+++ src/components/Map/ButtonGroup/UpperButtons/UpperButtonsPresenter.tsx
@@ -17,7 +17,6 @@ const UpperButtonsWrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: space-around;
-  /* align-items: center; */
 
   position: fixed;
   top: 10px;
