+++ src/components/Sidebar/SidebarContentMore/SidebarContentContainer.tsx
@@ -1,22 +1,14 @@
 import React, { useState } from 'react';
 import SidebarContentPresenter from './SidebarContentPresenter';
-import StylerPresenter from './StylerPresenter';
 import FeatureTypePresenter from './FeatureTypePresenter';
-import DetailTypePresenter from './DetailTypePresenter';
 
 function SidebarContentContainer(): React.ReactElement {
   const [featureName, setFeatureName] = useState('');
-  const [detailName, setDetailName] = useState(''); // (구역 채우기) 이런 형태
   const [styledFeatureList, setStyledFeatureList] = useState([]); // 추후 전역 상태로 변경
 
   const featureClickHandler = (name: string) => {
     if (name !== featureName) {
       setFeatureName(name);
-    }
-  };
-  const detailClickHandler = (name: string) => {
-    if (name !== detailName) {
-      setDetailName(name);
     }
   };
 
@@ -27,12 +19,6 @@ function SidebarContentContainer(): React.ReactElement {
         styledFeatureList={styledFeatureList}
         featureClickHandler={featureClickHandler}
       />
-      <DetailTypePresenter
-        featureName={featureName}
-        detailName={detailName}
-        detailClickHandler={detailClickHandler}
-      />
-      <StylerPresenter detailName={detailName} />
     </SidebarContentPresenter>
   );
 }
